# doc/api.yaml
openapi: 3.0.0
info:
  description: |-
    This concerns the "Web and Software Architecture" course to develop from scratch a web application that allows users to share photos of special moments.
    It allows users to like, comment on posts, ban other users, and so forth.

    Note: "Banning" and "setting a profile to private" are different. Setting a profile to public (private = false), means that any user except those whom he bans can see the user's profile, comment, like, etc without following the user. This is a feature much like "explore" in many social media. If a user wants only his followers to see his feeds except those banned, he should set it to private. Otherwise, even those who did not follow him and were not banned by him can see his feeds.
  title: Photo Sharing App
  version: 1.0.0
  contact:
    name: Yusupha Juwara
    url: https://github.com/YusuphaJuwara
    email: yusuphajuwara@gmail.com
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  parameters:
...
    userID:  #Arbitrary name for the definition that will be used to refer to it. Not necessarily the same as the parameter name.
      in: path
      name: user-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The user ID
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: uuid    # optional, arbitrary value for documentation purposes
  responses:
    NoContent204:
      description: Successful.
    BadRequest400:
      description: |-
        The request was not compliant with the documentation (eg. missing fields, etc).
        Further info in server logs.
    UnAuth401:
      description: |-
        Access token is missing or invalid.
        The request was not authorized. The user must be authorized first to perform this request.
    Forbidden403:
      description: |-
        Trying to access a resource that the user should not. This is forbidden action.
        E.g., 
        a banned user trying to get an info about the user who banned him, 
        or trying to delete a comment when the user is not the one who placed the comment or on whose post the comment was placed, etc.
        Further info in server logs.
    NotFound404:
      description: Not found.
    InternalServerError500:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  schemas:
    ID:
      description: The unique identifier. It is 36 characters containing 4 "-". 
      type: string
      format: uuid

      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 123e4567-e89b-12d3-a456-426655440000
    gender:
      description: The gender of the user
      type: string
      enum: 
        - Male
        - Female
      example: Male
      
      pattern: '^(Male|Female)$'
      minLength: 4
      maxLength: 6
    name:
      description: A string with a pattern that will be used elsewhere in the file such as profile name (not username for login).
      type: string
      pattern: '^(?=.*?[a-zA-Z]).{8,20}$' # 8 to 20 chars of at least 1 alph. No new line char
      minLength: 8
      maxLength: 20
      example: "Push Dollar $"
    username:
      description: |- 
        The username contains at least one alphabet and 7 or more other alphanumeric characters in [a-zA-Z0-9] to have a minLength of 8.
        This is only for login purposes. The profile name is the one used to display on the user's profile.
      type: string
      pattern: '^(?=.*?[a-zA-Z])[a-zA-Z0-9]{8,20}$'
      example: Maria2001
      minLength: 8
      maxLength: 20
    boolean:
      description: true or false
      type: boolean
      example: true
...
    date:
      description: The date.
      type: string
      format: date
      example: '2017-07-21'

      pattern: '^\d{4}-\d{2}-\d{2}$'
      minLength: 10   
      maxLength: 10
    message:
      description: The message as a comment or caption on a post.
      type: string
      pattern: '(?=.*?[a-zA-Z])[^]'  # Any char including new lines of at least one alphabet.
      minLength: 2
      maxLength: 1000
      example: "Wow, that's a beautifull family photo!!!"
...
    count:
      description: The total count of the object such as total number of likes, total number of comments, etc.
      type: integer
      minimum: 0
      example: 1000
    ids-array:
      description: array of uuid's.
      type: array
      minItems: 0
      maxItems: 1000
      items:  
        $ref: '#/components/schemas/ID'
    user:
      type: object
      description: The user-associated details.
      properties:
        user-id:  
          $ref: '#/components/schemas/ID'
        profile-name:   
          $ref: '#/components/schemas/name'
        gender:   
          $ref: "#/components/schemas/gender"
        birth-date:   
          $ref: '#/components/schemas/date'
        profile-message:   
          $ref: '#/components/schemas/message'
        profile-photo-id:   
          $ref: '#/components/schemas/ID'
        user-post-ids:  
            $ref: '#/components/schemas/ids-array'
        post-count:  
          $ref: '#/components/schemas/count'
        follower-count:  
          $ref: '#/components/schemas/count'
        following-count:   
          $ref: '#/components/schemas/count'
...
security:
  - bearerToken: []
...
paths:
  /users/:
    get:
      tags: [User]
      operationId: getAllUsers
      summary: Get an array of information of all users
      description: |-
        For each user in our database, this gets it's information. 
        But the user trying to get users' info cannot get those users who banned him or those whom he did not "follow" and set their profiles to private.
      responses:
        "200":
          description: Array of users
          content:
            application/json:
              schema:
                description: Array of users' info
                type: array
                minItems: 0
                maxItems: 1000
                items:  
                  $ref: "#/components/schemas/user"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
...
  /users/{user-id}/private-profile:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [User]
      operationId: getPrivate
      summary: Whether or not the profile is private.
      description: |-
        See if the user profile is set to private or not.
      responses:
        "200":
          description: True or false.
          content:
            application/json:
              schema:   
                $ref: "#/components/schemas/boolean"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    put:
      tags: [User]
      operationId: setPrivate
      summary: Sets the user's profile to private.
      description: |-
        Sets the user's profile to private if not set already.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    delete:
      tags: [User]
      operationId: setPublic
      summary: Sets the user's profile to public.
      description: |-
        Sets the user's profile to public if not set already.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [User]
      operationId: getUserProfile
      summary: Get an information of a user.
      description: |-
        Get these information about the given user:
        user-id, profile-name, profile-message, gender, birth-date, profile-photo-id, array of post IDs that the user posted sorted according to last-post-first, post-count, follower-count, following-count.
      responses:
        "200":
          description: user's information.
          content:
            application/json:
              schema:   
                $ref: "#/components/schemas/user"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    patch:
      tags: [User]
      operationId: updateUserProfile
      summary: Update the user profile details.
      description: |-
        Update at least one of these: profile-name, profile-message, gender, birth-date.
        If this leads to the creation of any of the fields on the user profile, 201 is returned.
      requestBody:
        required: true
        content:
          application/json:
            schema:  
              description: At least one property must be set.
              type: object
              properties:   
                profile-name:
                  $ref: '#/components/schemas/name'
                profile-message:
                  $ref: '#/components/schemas/message'
                gender:
                  $ref: '#/components/schemas/gender'
                birth-date:
                  $ref: '#/components/schemas/date'
              minProperties: 1   
              additionalProperties: false 
      responses:
        "201":  # --
          description: Successfully created some fields that did not exist. Send back the possible modified fields.
          content:
            application/json:
              schema:  
                description: Schema structure for fields that are possible to modify.
                type: object
                properties:   
                  profile-name:
                    $ref: '#/components/schemas/name'
                  profile-message:
                    $ref: '#/components/schemas/message'
                  gender:
                    $ref: '#/components/schemas/gender'
                  birth-date:
                    $ref: '#/components/schemas/date'
        "204":
          $ref: '#/components/responses/NoContent204'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    delete:
      tags: [User]
      operationId: deleteUser
      summary: Deletes the account of the user permanently.
      description: |-
        Deletes the user together with all its information.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/username:
    parameters:
      - $ref: "#/components/parameters/userID"
    put:
      tags: [User]
      operationId: setMyUserName
      summary: Modify user profile name.
      description: |-
        Replaces the user profile name to the specified name in the request body.
      requestBody:
        required: true
        content:
          application/json:
            schema:   
              $ref: '#/components/schemas/username'
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
...